standards and nat rankings

+++ src/types/standards.rs
@@ -4,6 +4,6 @@ use serde::Deserialize;
 #[serde(rename_all = "camelCase")]
 pub struct Standards {
     pub weight_class: String,
-    pub standard_a: i32,
-    pub standard_b: i32,
+    pub standard_a: f64,
+    pub standard_b: f64,
 }
